backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
Pillow>=10.0.0
psutil>=5.9.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.8.0
python-multipart>=0.0.6
